# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ApiErrorData": {
				"description": "에러 응답 추가 데이터",
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"anyOf": [
						{
							"type": "string"
						},
						{
							"type": "number",
							"format": "double"
						},
						{
							"type": "boolean"
						}
					],
					"nullable": true
				}
			},
			"ApiErrorResponseBody": {
				"description": "API 실패 응답 에러 상세 정보",
				"properties": {
					"errorCode": {
						"type": "string",
						"description": "에러 코드"
					},
					"message": {
						"type": "string",
						"description": "클라이언트 전달 에러 메시지"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ApiErrorData"
							}
						],
						"nullable": true,
						"description": "에러 관련 추가 데이터"
					}
				},
				"required": [
					"errorCode",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiFailResponse": {
				"description": "API 실패 응답 공통 포맷",
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"FAIL"
						],
						"nullable": false,
						"description": "요청 처리 결과"
					},
					"error": {
						"$ref": "#/components/schemas/ApiErrorResponseBody",
						"description": "실패 원인 상세 정보"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true,
						"description": "실패 응답 null 값"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ReviewResponse": {
				"description": "리뷰 정보",
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "리뷰 ID"
					},
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "리뷰 등록 가게 ID"
					},
					"memberId": {
						"type": "number",
						"format": "double",
						"description": "리뷰 작성 회원 ID"
					},
					"rating": {
						"type": "number",
						"format": "double",
						"description": "평점"
					},
					"content": {
						"type": "string",
						"description": "리뷰 내용"
					},
					"createdAt": {
						"type": "string",
						"description": "리뷰 생성 시각"
					}
				},
				"required": [
					"id",
					"storeId",
					"memberId",
					"rating",
					"content",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiSuccessResponse_ReviewResponse_": {
				"description": "API 성공 응답 공통 포맷",
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false,
						"description": "요청 처리 결과"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true,
						"description": "성공 응답 null 값"
					},
					"data": {
						"$ref": "#/components/schemas/ReviewResponse",
						"description": "API별 성공 응답 데이터"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateReviewRequest": {
				"description": "리뷰 생성 요청",
				"properties": {
					"rating": {
						"type": "number",
						"format": "double",
						"description": "평점, 1부터 5까지"
					},
					"content": {
						"type": "string",
						"description": "리뷰 내용"
					}
				},
				"required": [
					"rating",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StoreResponse": {
				"description": "가게 정보",
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "가게 ID"
					},
					"regionId": {
						"type": "number",
						"format": "double",
						"description": "가게 소속 지역 ID"
					},
					"name": {
						"type": "string",
						"description": "가게 이름"
					},
					"address": {
						"type": "string",
						"description": "가게 주소"
					}
				},
				"required": [
					"id",
					"regionId",
					"name",
					"address"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"MemberMissionResponse": {
				"description": "회원 미션 도전 정보",
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "회원 미션 ID"
					},
					"memberId": {
						"type": "number",
						"format": "double",
						"description": "회원 ID"
					},
					"missionId": {
						"type": "number",
						"format": "double",
						"description": "미션 ID"
					},
					"status": {
						"type": "string",
						"description": "미션 진행 상태"
					},
					"challengedAt": {
						"type": "string",
						"description": "미션 도전 시각"
					}
				},
				"required": [
					"id",
					"memberId",
					"missionId",
					"status",
					"challengedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiSuccessResponse_MemberMissionResponse_": {
				"description": "API 성공 응답 공통 포맷",
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false,
						"description": "요청 처리 결과"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true,
						"description": "성공 응답 null 값"
					},
					"data": {
						"$ref": "#/components/schemas/MemberMissionResponse",
						"description": "API별 성공 응답 데이터"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MemberResponse": {
				"description": "회원 정보",
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "회원 ID"
					},
					"name": {
						"type": "string",
						"description": "회원 이름"
					},
					"email": {
						"type": "string",
						"nullable": true,
						"description": "회원 이메일"
					},
					"phone": {
						"type": "string",
						"nullable": true,
						"description": "회원 전화번호"
					},
					"birthday": {
						"type": "string",
						"nullable": true,
						"description": "회원 생일"
					},
					"authProvider": {
						"type": "string",
						"description": "최초 가입 방식"
					}
				},
				"required": [
					"id",
					"name",
					"authProvider"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthTokenResponse": {
				"description": "로그인 토큰 응답",
				"properties": {
					"accessToken": {
						"type": "string",
						"description": "JWT Access Token"
					},
					"member": {
						"$ref": "#/components/schemas/MemberResponse",
						"description": "로그인한 회원 정보"
					}
				},
				"required": [
					"accessToken",
					"member"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiSuccessResponse_AuthTokenResponse_": {
				"description": "API 성공 응답 공통 포맷",
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false,
						"description": "요청 처리 결과"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true,
						"description": "성공 응답 null 값"
					},
					"data": {
						"$ref": "#/components/schemas/AuthTokenResponse",
						"description": "API별 성공 응답 데이터"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SignUpRequest": {
				"description": "회원가입 요청",
				"properties": {
					"email": {
						"type": "string",
						"description": "회원 이메일"
					},
					"password": {
						"type": "string",
						"description": "회원 비밀번호"
					},
					"name": {
						"type": "string",
						"description": "회원 이름"
					},
					"phone": {
						"type": "string",
						"description": "회원 전화번호"
					},
					"birthday": {
						"type": "string",
						"description": "회원 생일"
					}
				},
				"required": [
					"email",
					"password",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"LoginRequest": {
				"description": "로그인 요청",
				"properties": {
					"email": {
						"type": "string",
						"description": "회원 이메일"
					},
					"password": {
						"type": "string",
						"description": "회원 비밀번호"
					}
				},
				"required": [
					"email",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiSuccessResponse_MemberResponse_": {
				"description": "API 성공 응답 공통 포맷",
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false,
						"description": "요청 처리 결과"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true,
						"description": "성공 응답 null 값"
					},
					"data": {
						"$ref": "#/components/schemas/MemberResponse",
						"description": "API별 성공 응답 데이터"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateMyInfoRequest": {
				"description": "내 정보 수정 요청",
				"properties": {
					"name": {
						"type": "string",
						"description": "회원 이름"
					},
					"phone": {
						"type": "string",
						"description": "회원 전화번호"
					},
					"birthday": {
						"type": "string",
						"description": "회원 생일"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"ReviewWithStoreResponse": {
				"description": "가게 정보를 포함한 리뷰 정보",
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "리뷰 ID"
					},
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "리뷰 등록 가게 ID"
					},
					"memberId": {
						"type": "number",
						"format": "double",
						"description": "리뷰 작성 회원 ID"
					},
					"rating": {
						"type": "number",
						"format": "double",
						"description": "평점"
					},
					"content": {
						"type": "string",
						"description": "리뷰 내용"
					},
					"createdAt": {
						"type": "string",
						"description": "리뷰 생성 시각"
					},
					"store": {
						"$ref": "#/components/schemas/StoreResponse",
						"description": "리뷰 등록 가게 정보"
					}
				},
				"required": [
					"id",
					"storeId",
					"memberId",
					"rating",
					"content",
					"createdAt",
					"store"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiSuccessResponse_ReviewWithStoreResponse-Array_": {
				"description": "API 성공 응답 공통 포맷",
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false,
						"description": "요청 처리 결과"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true,
						"description": "성공 응답 null 값"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/ReviewWithStoreResponse"
						},
						"type": "array",
						"description": "API별 성공 응답 데이터"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MissionWithStoreResponse": {
				"description": "가게 정보를 포함한 미션 정보",
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "미션 ID"
					},
					"storeId": {
						"type": "number",
						"format": "double",
						"description": "미션 등록 가게 ID"
					},
					"title": {
						"type": "string",
						"description": "미션 제목"
					},
					"reward": {
						"type": "number",
						"format": "double",
						"description": "미션 보상 포인트"
					},
					"deadline": {
						"type": "string",
						"description": "미션 마감일"
					},
					"createdAt": {
						"type": "string",
						"description": "미션 생성 시각"
					},
					"store": {
						"$ref": "#/components/schemas/StoreResponse",
						"description": "미션 등록 가게 정보"
					}
				},
				"required": [
					"id",
					"storeId",
					"title",
					"reward",
					"deadline",
					"createdAt",
					"store"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MemberMissionWithMissionResponse": {
				"description": "미션 상세 정보를 포함한 회원 미션 도전 정보",
				"properties": {
					"id": {
						"type": "number",
						"format": "double",
						"description": "회원 미션 ID"
					},
					"memberId": {
						"type": "number",
						"format": "double",
						"description": "회원 ID"
					},
					"missionId": {
						"type": "number",
						"format": "double",
						"description": "미션 ID"
					},
					"status": {
						"type": "string",
						"description": "미션 진행 상태"
					},
					"challengedAt": {
						"type": "string",
						"description": "미션 도전 시각"
					},
					"mission": {
						"$ref": "#/components/schemas/MissionWithStoreResponse",
						"description": "도전 중인 미션 상세 정보"
					}
				},
				"required": [
					"id",
					"memberId",
					"missionId",
					"status",
					"challengedAt",
					"mission"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiSuccessResponse_MemberMissionWithMissionResponse-Array_": {
				"description": "API 성공 응답 공통 포맷",
				"properties": {
					"resultType": {
						"type": "string",
						"enum": [
							"SUCCESS"
						],
						"nullable": false,
						"description": "요청 처리 결과"
					},
					"error": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true,
						"description": "성공 응답 null 값"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/MemberMissionWithMissionResponse"
						},
						"type": "array",
						"description": "API별 성공 응답 데이터"
					}
				},
				"required": [
					"resultType",
					"error",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "umc-node-simple-api",
		"version": "1.0.0",
		"description": "Simple Express API server with SQLite dummy data.",
		"contact": {}
	},
	"paths": {
...
		"/stores/{storeId}/reviews": {
			"post": {
				"operationId": "CreateReview",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_ReviewResponse_"
								}
							}
						}
					},
					"400": {
						"description": "storeId, rating, content가 유효하지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					},
					"401": {
						"description": "로그인이 필요하거나 토큰이 유효하지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 가게인 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					}
				},
				"description": "특정 가게 리뷰 작성",
				"tags": [
					"Stores"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "가게 ID",
						"in": "path",
						"name": "storeId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "리뷰 생성 요청 body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateReviewRequest",
								"description": "리뷰 생성 요청 body"
							}
						}
					}
				}
			}
		},
...
		"/missions/{missionId}/challenges": {
			"post": {
				"operationId": "ChallengeMission",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_MemberMissionResponse_"
								}
							}
						}
					},
					"400": {
						"description": "missionId가 유효하지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					},
					"401": {
						"description": "로그인이 필요하거나 토큰이 유효하지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					},
					"404": {
						"description": "존재하지 않는 미션인 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					},
					"409": {
						"description": "이미 도전 중인 미션인 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					}
				},
				"description": "특정 미션 도전",
				"tags": [
					"Missions"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "미션 ID",
						"in": "path",
						"name": "missionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/missions/{missionId}/challenges/complete": {
			"patch": {
				"operationId": "CompleteChallenge",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_MemberMissionResponse_"
								}
							}
						}
					},
					"400": {
						"description": "missionId가 유효하지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					},
					"401": {
						"description": "로그인이 필요하거나 토큰이 유효하지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					},
					"404": {
						"description": "진행 중인 미션이 아닌 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					}
				},
				"description": "진행 중인 미션 완료 처리",
				"tags": [
					"Missions"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "미션 ID",
						"in": "path",
						"name": "missionId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/auth/signup": {
			"post": {
				"operationId": "SignUp",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_AuthTokenResponse_"
								}
							}
						}
					},
					"400": {
						"description": "email, password, name이 유효하지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					},
					"409": {
						"description": "다른 로그인 방식으로 가입된 이메일인 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					}
				},
				"description": "이메일 회원가입",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "회원가입 요청 body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SignUpRequest",
								"description": "회원가입 요청 body"
							}
						}
					}
				}
			}
		},
...
		"/auth/google/callback": {
			"get": {
				"operationId": "GoogleCallback",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_AuthTokenResponse_"
								}
							}
						}
					},
					"401": {
						"description": "Google 로그인이 실패한 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					},
					"409": {
						"description": "다른 로그인 방식으로 가입된 이메일인 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					}
				},
				"description": "Google 로그인 콜백",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": []
			}
		},
		"/auth/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_AuthTokenResponse_"
								}
							}
						}
					},
					"400": {
						"description": "email, password가 유효하지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					},
					"404": {
						"description": "이메일 또는 비밀번호가 올바르지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					}
				},
				"description": "이메일 로그인",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "로그인 요청 body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LoginRequest",
								"description": "로그인 요청 body"
							}
						}
					}
				}
			}
		},
		"/members/me": {
			"get": {
				"operationId": "GetMyInfo",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_MemberResponse_"
								}
							}
						}
					},
					"401": {
						"description": "로그인이 필요하거나 토큰이 유효하지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					}
				},
				"description": "현재 로그인한 회원 정보 조회",
				"tags": [
					"Members"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			},
			"patch": {
				"operationId": "UpdateMyInfo",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_MemberResponse_"
								}
							}
						}
					},
					"400": {
						"description": "수정할 회원 정보가 없는 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					},
					"401": {
						"description": "로그인이 필요하거나 토큰이 유효하지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					}
				},
				"description": "현재 로그인한 회원 정보 수정",
				"tags": [
					"Members"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "내 정보 수정 요청 body",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateMyInfoRequest",
								"description": "내 정보 수정 요청 body"
							}
						}
					}
				}
			}
		},
		"/members/me/reviews": {
			"get": {
				"operationId": "GetMyReviews",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_ReviewWithStoreResponse-Array_"
								}
							}
						}
					},
					"400": {
						"description": "limit, cursor가 유효하지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					},
					"401": {
						"description": "로그인이 필요하거나 토큰이 유효하지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					}
				},
				"description": "현재 로그인한 회원 리뷰 목록 조회",
				"tags": [
					"Members"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "한 번에 조회할 개수, 기본값 10, 최대 50",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "이전 페이지 마지막 리뷰 ID",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/members/me/missions": {
			"get": {
				"operationId": "GetMyInProgressMissions",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiSuccessResponse_MemberMissionWithMissionResponse-Array_"
								}
							}
						}
					},
					"400": {
						"description": "limit, cursor가 유효하지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					},
					"401": {
						"description": "로그인이 필요하거나 토큰이 유효하지 않은 경우",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiFailResponse"
								}
							}
						}
					}
				},
				"description": "현재 로그인한 회원 진행 중 미션 목록 조회",
				"tags": [
					"Members"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "한 번에 조회할 개수, 기본값 10, 최대 50",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "이전 페이지 마지막 회원 미션 ID",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}